implemented pollen-collector refactor.
